完善单指计算器算法点击功能

+++ app/src/main/java/com/example/testdemo/keyboard/datastruct/HandMark.java
@@ -17,7 +17,7 @@ public class HandMark {
 
     public Position[] jointPoint =new Position[21];
     public int ID;
-    public double timestamp=0.0;
+    public long timestamp=0;
     public boolean historyMoveSign =false;
     public boolean historyPushSign =false;
     public boolean historyFOnKSign =false;
